still keep the req.txt for init in the prerequisites.py
